src/gaze_tracker/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.11.0=py36h5fab9bb_0
astor=0.8.1=pyh9f0ad1d_0
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h36c2ea0_0
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.1=py_0
cairo=1.16.0=h7979940_1007
certifi=2020.12.5=py36h5fab9bb_1
cloudpickle=1.6.0=py_0
cudatoolkit=11.2.1=h8204236_8
cycler=0.10.0=py_2
cython=0.29.21=py36hc4f0c31_2
cytoolz=0.11.0=py36h8f6f2f9_3
dask-core=2021.2.0=pyhd8ed1ab_0
dbus=1.13.6=hfdff14a_1
decorator=4.4.2=py_0
easydict=1.9=py_0
expat=2.2.10=h9c3ff4c_0
ffmpeg=4.3.1=hca11adc_2
fontconfig=2.13.1=hba837de_1004
freetype=2.10.4=h0708190_1
gast=0.2.2=py36_0
geos=3.9.1=h9c3ff4c_2
gettext=0.19.8.1=h0b5b191_1005
glib=2.66.7=h9c3ff4c_1
glib-tools=2.66.7=h9c3ff4c_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-pasta=0.2.0=pyh8c360ce_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.36.0=py36h8e87921_0
gst-plugins-base=1.18.3=h04508c2_0
gstreamer=1.18.3=h3560a44_0
h5py=2.10.0=nompi_py36h4510012_106
harfbuzz=2.7.4=h5cf4720_0
hdf5=1.10.6=nompi_h6a2412b_1114
icu=68.1=h58526e2_0
imagecodecs-lite=2019.12.3=py36h92226af_3
imageio=2.9.0=py_0
imgaug=0.4.0=py_1
importlib-metadata=3.7.0=py36h5fab9bb_0
jasper=1.900.1=h07fcdf6_1006
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
keras=2.3.1=py36_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.3.1=py36h605e78d_1
krb5=1.17.2=h926e7f8_0
lame=3.100=h7f98852_1001
lcms2=2.12=hddcbb42_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libclang=11.0.1=default_ha53f305_1
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=hcdb4288_3
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=9.3.0=hff62375_18
libgfortran5=9.3.0=hff62375_18
libglib=2.66.7=h3e27bee_1
libgomp=9.3.0=h2828fa1_18
libgpuarray=0.7.6=h7f98852_1003
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_openblas
liblapacke=3.9.0=8_openblas
libllvm10=10.0.1=he513fc3_3
libllvm11=11.0.1=hf817b99_0
libnghttp2=1.43.0=h812cca2_0
libopenblas=0.3.12=pthreads_h4812303_1
libopencv=4.5.1=py36h1da8354_0
libpng=1.6.37=h21135ba_2
libpq=13.1=hfd2b0eb_1
libprotobuf=3.15.3=h780b84a_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h6de172a_18
libtiff=4.2.0=hdc55705_0
libuuid=2.32.1=h7f98852_1000
libwebp-base=1.2.0=h7f98852_0
libxcb=1.13=h7f98852_1003
libxkbcommon=1.0.3=he3ba5ed_0
libxml2=2.9.10=h72842e0_3
llvmlite=0.35.0=py36h05121d2_1
lz4-c=1.9.3=h9c3ff4c_0
mako=1.1.4=pyh44b312d_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=1.1.1=py36h8f6f2f9_3
matplotlib-base=3.3.4=py36hd391965_0
mtcnn=0.1.0=py_0
mysql-common=8.0.23=ha770c72_1
mysql-libs=8.0.23=h935591d_1
ncurses=6.2=h58526e2_4
nettle=3.6=he412f7d_0
networkx=2.5=py_0
nspr=4.29=h9c3ff4c_1
nss=3.62=hb5efdd6_0
numba=0.52.0=py36h284efc9_0
numpy=1.19.5=py36h2aa4a07_1
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.1=py36h5fab9bb_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1k=h7f98852_0
#optunity=1.1.1=py36h7da2bbe_0
pcre=8.44=he1b5a44_0
pillow=8.1.0=py36ha6010c0_2
pip=21.0.1=pyhd8ed1ab_0
pixman=0.40.0=h36c2ea0_0
protobuf=3.15.3=py36hc4f0c31_0
pthread-stubs=0.4=h36c2ea0_1001
py-opencv=4.5.1=py36he448a4c_0
pygpu=0.7.6=py36h92226af_1002
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.6.12=hffdb5ce_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.6=1_cp36m
pywavelets=1.1.1=py36h92226af_3
pyyaml=5.4.1=py36h8f6f2f9_0
qt=5.12.9=hda022c4_4
readline=8.0=he28a2e2_2
scikit-image=0.17.2=py36h284efc9_4
scikit-learn=0.24.1=py36he4fde30_0
scipy=1.5.3=py36h9e8f40b_0
setuptools=49.6.0=py36h5fab9bb_3
shapely=1.7.1=py36h93b233e_4
six=1.15.0=pyh9f0ad1d_0
sqlite=3.34.0=h74cdb3f_0
tensorboard=1.14.0=py36_0
tensorflow=1.14.0=hc3e5e64_0
tensorflow-base=1.14.0=py36hc3e5e64_0
tensorflow-estimator=1.14.0=py36h5ca1d4c_0
termcolor=1.1.0=py_2
theano=0.9.0=py36_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.6.3=py_0
tk=8.6.10=h21135ba_1
toolz=0.11.1=py_0
tornado=6.1=py36h8f6f2f9_1
typing_extensions=3.7.4.3=py_0
werkzeug=1.0.1=pyh9f0ad1d_0
wheel=0.36.2=pyhd3deb0d_0
wrapt=1.12.1=py36h8f6f2f9_3
x264=1!161.3030=h7f98852_0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h516909a_0
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.8=ha95c52a_1
